requirements.txt
streamlit==1.44.0
pandas==2.2.3
plotly==5.24.1
vaderSentiment==3.3.2
requests==2.32.3
google-api-python-client>=2.157.0
duckduckgo-search
ddgs
googlesearch-python>=1.2.3
beautifulsoup4>=4.12
nltk>=3.8
scikit-learn>=1.4
